split out extension stuff from pydevd_plugin_utils to avoid circular imports
